+++ yash-syntax/src/parser/lex/core.rs
@@ -36,6 +36,7 @@ use std::future::Future;
 use std::num::NonZeroU64;
 use std::ops::Deref;
 use std::ops::DerefMut;
+use std::ops::Range;
 use std::pin::Pin;
 use std::rc::Rc;
 use std::slice::SliceIndex;
@@ -278,6 +279,26 @@ impl<'a> LexerCore<'a> {
         self.source[i].iter().map(|c| c.value).collect()
     }
 
+    /// Returns a location for a given range of the source code.
+    #[must_use]
+    fn location_range(&self, range: Range<usize>) -> Location {
+        if range.start == self.source.len() {
+            if let InputState::EndOfInput(ref location) = self.state {
+                return location.clone();
+            }
+        }
+        let start = &self.peek_char_at(range.start).location;
+        let code = start.code.clone();
+        let end = range
+            .map(|index| &self.peek_char_at(index).location)
+            .take_while(|location| location.code == code)
+            .last()
+            .map(|location| location.range.end)
+            .unwrap_or(start.range.start);
+        let range = start.range.start..end;
+        Location { code, range }
+    }
+
     /// Performs alias substitution.
     fn substitute_alias(&mut self, begin: usize, alias: &Rc<Alias>) {
         let end = self.index;
@@ -455,6 +476,9 @@ impl<'a> Lexer<'a> {
     /// and a newline are silently skipped before returning the next character.
     /// Call [`disable_line_continuation`](Self::disable_line_continuation) to
     /// switch off line continuation recognition.
+    ///
+    /// This function requires a mutable reference to `self` since it may need
+    /// to read the next line if needed.
     pub async fn peek_char(&mut self) -> Result<Option<char>> {
         while self.line_continuation().await? {}
 
@@ -469,8 +493,8 @@ impl<'a> Lexer<'a> {
     /// If there is no more character (that is, it is the end of input), an imaginary location
     /// is returned that would be returned if a character existed.
     ///
-    /// This function requires a mutable reference to `self` since it may need to read a next
-    /// line if it is not yet read.
+    /// This function requires a mutable reference to `self` since it needs to
+    /// [peek](Self::peek_char) the next character.
     pub async fn location(&mut self) -> Result<&Location> {
         self.core.peek_char().await.map(|p| p.location())
     }
@@ -597,6 +621,27 @@ impl<'a> Lexer<'a> {
         I: SliceIndex<[SourceChar], Output = [SourceChar]>,
     {
         self.core.source_string(i)
+    }
+
+    /// Returns a location for a given range of the source code.
+    ///
+    /// All the characters in the range must have been
+    /// [consume](Self::consume_char)d. If the range refers to an unconsumed
+    /// character, this function will panic!
+    ///
+    /// If the characters are from more than one [`Code`] fragment, the location
+    /// will only cover the initial portion of the range sharing the same
+    /// `Code`.
+    ///
+    /// # Panics
+    ///
+    /// This function will panic if the range refers to an unconsumed character.
+    ///
+    /// If the start index of the range is the end of input, it must have been
+    /// peeked and the range must be empty, or the function will panic.
+    #[must_use]
+    pub fn location_range(&self, range: Range<usize>) -> Location {
+        self.core.location_range(range)
     }
 
     /// Performs alias substitution right before the current position.
@@ -1323,6 +1368,101 @@ mod tests {
             unreachable!("unexpected call to the decider function: argument={}", c)
         }));
         assert_eq!(r, Ok(None));
+    }
+
+    #[test]
+    fn lexer_location_range_with_empty_range() {
+        let mut lexer = Lexer::from_memory("", Source::Unknown);
+        block_on(lexer.peek_char()).unwrap();
+        let location = lexer.location_range(0..0);
+        assert_eq!(*location.code.value.borrow(), "");
+        assert_eq!(location.code.start_line_number.get(), 1);
+        assert_eq!(location.code.source, Source::Unknown);
+        assert_eq!(location.range, 0..0);
+    }
+
+    #[test]
+    fn lexer_location_range_with_nonempty_range() {
+        block_on(async {
+            let mut lexer = Lexer::from_memory("cat foo", Source::Stdin);
+            for _ in 0..4 {
+                lexer.peek_char().await.unwrap();
+                lexer.consume_char();
+            }
+            lexer.peek_char().await.unwrap();
+
+            let location = lexer.location_range(1..4);
+            assert_eq!(*location.code.value.borrow(), "cat foo");
+            assert_eq!(location.code.start_line_number.get(), 1);
+            assert_eq!(location.code.source, Source::Stdin);
+            assert_eq!(location.range, 1..4);
+        })
+    }
+
+    #[test]
+    fn lexer_location_range_with_range_starting_at_end() {
+        block_on(async {
+            let mut lexer = Lexer::from_memory("cat", Source::Stdin);
+            for _ in 0..3 {
+                lexer.peek_char().await.unwrap();
+                lexer.consume_char();
+            }
+            lexer.peek_char().await.unwrap();
+
+            let location = lexer.location_range(3..3);
+            assert_eq!(*location.code.value.borrow(), "cat");
+            assert_eq!(location.code.start_line_number.get(), 1);
+            assert_eq!(location.code.source, Source::Stdin);
+            assert_eq!(location.range, 3..3);
+        })
+    }
+
+    #[test]
+    #[should_panic]
+    fn lexer_location_range_with_unconsumed_code() {
+        let lexer = Lexer::from_memory("echo ok", Source::Unknown);
+        let _ = lexer.location_range(0..0);
+    }
+
+    #[test]
+    #[should_panic(expected = "The index 1 must not be larger than the current index 0")]
+    fn lexer_location_range_with_range_out_of_bounds() {
+        let lexer = Lexer::from_memory("", Source::Unknown);
+        let _ = lexer.location_range(1..2);
+    }
+
+    #[test]
+    fn lexer_location_range_with_alias_substitution() {
+        block_on(async {
+            let mut lexer = Lexer::from_memory(" a;", Source::Unknown);
+            let alias_def = Rc::new(Alias {
+                name: "a".to_string(),
+                replacement: "abc".to_string(),
+                global: false,
+                origin: Location::dummy("dummy"),
+            });
+            for _ in 0..2 {
+                lexer.peek_char().await.unwrap();
+                lexer.consume_char();
+            }
+            lexer.substitute_alias(1, &alias_def);
+            for _ in 1..5 {
+                lexer.peek_char().await.unwrap();
+                lexer.consume_char();
+            }
+
+            let location = lexer.location_range(2..5);
+            assert_eq!(*location.code.value.borrow(), "abc");
+            assert_eq!(location.code.start_line_number.get(), 1);
+            assert_matches!(&location.code.source, Source::Alias { original, alias } => {
+                assert_eq!(*original.code.value.borrow(), " a;");
+                assert_eq!(original.code.start_line_number.get(), 1);
+                assert_eq!(original.code.source, Source::Unknown);
+                assert_eq!(original.range, 1..2);
+                assert_eq!(alias, &alias_def);
+            });
+            assert_eq!(location.range, 1..3);
+        })
     }
 
     #[test]